feat(build-wms-layer): Preserve query params
Before, it was not possible to add query params to server urls in the
config file. Now the query params are preserved and merged with the
other paramters.

This is implemented by converting the raw url (containing the query
params) into an URL object, extracting the query params and merging it
with the other config variables. The query params are added last, this
allows for overwriting the default config.