feat: Load multiple configs during page load
When multiple configs are specified, each one is being loaded. They are
combined. The root object in the data structure is created in the
getConfig function. The config data structure is now even simpler
because at each level children are in the `children` property, where
before, the root level had `layers`